# docs/_static/files/swagger.json
{
    "swagger": "2.0",
    "basePath": "\/api",
    "paths": {
...
        "\/browse\/{vendor}": {
            "parameters": [
                {
                    "name": "vendor",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Get the known products for a vendor",
                "description": "Get the known products for a vendor.",
                "operationId": "get_vendor_products",
                "tags": [
                    "browse"
                ]
            }
        },
...
        "\/product\/": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#\/definitions\/ProductsList"
                            }
                        }
                    }
                },
                "summary": "List all products",
                "operationId": "list_products",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "Number of the page.",
                        "default": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "type": "integer",
                        "description": "Maximum number of elements to return.",
                        "default": 1000
                    },
                    {
                        "name": "uuid",
                        "in": "query",
                        "type": "string",
                        "description": "UUID of the product."
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "type": "string",
                        "description": "The name of the product."
                    },
                    {
                        "name": "organization_name",
                        "in": "query",
                        "type": "string",
                        "description": "The name of the organization related to the product."
                    },
                    {
                        "name": "organization_id",
                        "in": "query",
                        "type": "string",
                        "description": "The id of the organization related to the product."
                    },
                    {
                        "name": "organization_uuid",
                        "in": "query",
                        "type": "string",
                        "description": "The uuid of the organization related to the product."
                    },
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "tags": [
                    "product"
                ]
            }
        },
...
        "\/vulnerability\/": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success."
                    },
                    "400": {
                        "description": "JSON validation failed."
                    },
                    "403": {
                        "description": "Reporter permission required."
                    },
                    "422": {
                        "description": "Not possible to edit a vulnerability from the requested source."
                    }
                },
                "summary": "Endpoint for creating and editing vulnerabilities in the local source",
                "description": "Create a vulnerability with the CVE version 5 format.",
                "operationId": "post_vulnerabilities_list",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "tags": [
                    "vulnerability"
                ]
            }
        },
...
        "\/vulnerability\/last\/{number}": {
            "parameters": [
                {
                    "name": "number",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Retrieve the latest vulnerabilities, with optional filters for source and number of results",
                "description": "Get the last vulnerabilities\nSupports light mode for minimal data and provides backward-compatible routes.",
                "operationId": "get_last",
                "tags": [
                    "vulnerability"
                ]
            }
        },
        "\/vulnerability\/last\/{source}": {
            "parameters": [
                {
                    "name": "source",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Retrieve the latest vulnerabilities, with optional filters for source and number of results",
                "description": "Get the last vulnerabilities\nSupports light mode for minimal data and provides backward-compatible routes.",
                "operationId": "get_last",
                "tags": [
                    "vulnerability"
                ]
            }
        },
        "\/vulnerability\/last\/{source}\/{number}": {
            "parameters": [
                {
                    "name": "source",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "number",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Retrieve the latest vulnerabilities, with optional filters for source and number of results",
                "description": "Get the last vulnerabilities\nSupports light mode for minimal data and provides backward-compatible routes.",
                "operationId": "get_last",
                "tags": [
                    "vulnerability"
                ]
            }
        },
...
        "\/vulnerability\/recent\/{date}": {
            "parameters": [
                {
                    "name": "date",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Retrieve vulnerabilities reported after a specified date, with optional filters for source and number of results",
                "description": "Retrieve vulnerabilities reported after a specified date, with optional filters for source and number of results. Defaults to the last 24 hours if no date is provided. Supports light mode for minimal data.\nDefaults to the last 24 hours if no date is provided. Supports light mode for minimal data.",
                "operationId": "get_recent",
                "tags": [
                    "vulnerability"
                ]
            }
        },
...
        "\/vulnerability\/search\/{vendor}\/{product}": {
            "parameters": [
                {
                    "name": "vendor",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "product",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Returns a list of vulnerabilities related to the vendor and product",
                "description": "Returns a list of vulnerabilities related to the product.\nOptionnaly filter vulnerabilities published or updated after the specified date (format: YYYY-MM-DD).",
                "operationId": "get_vendor_product_vulnerabilities",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "Number of the page.",
                        "default": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "type": "integer",
                        "description": "Maximum number of elements to return.",
                        "default": 10
                    },
                    {
                        "name": "since",
                        "in": "query",
                        "type": "string",
                        "description": "Filter vulnerabilities published or updated after the specified date (format: YYYY-MM-DD)."
                    }
                ],
                "tags": [
                    "vulnerability"
                ]
            }
        },
        "\/vulnerability\/{vulnerability_id}": {
            "parameters": [
                {
                    "name": "vulnerability_id",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "delete": {
                "responses": {
                    "204": {
                        "description": "Success."
                    },
                    "403": {
                        "description": "Admin permission required."
                    },
                    "422": {
                        "description": "Not possible to edit a vulnerability from the requested source."
                    }
                },
                "summary": "Endpoint for deleting a vulnerability",
                "description": "Delete a vulnerability from the local source.\nWe only accept to delete vulnerabilities from the local source.",
                "operationId": "delete_vulnerability",
                "tags": [
                    "vulnerability"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Get a vulnerability with its id",
                "description": "Get a vulnerability.",
                "operationId": "get_vulnerability",
                "parameters": [
                    {
                        "name": "with_meta",
                        "in": "query",
                        "type": "boolean",
                        "description": "Include metada.",
                        "default": false
                    },
                    {
                        "name": "with_linked",
                        "in": "query",
                        "type": "boolean",
                        "description": "Include the linked vulnerabilities.",
                        "default": false
                    },
                    {
                        "name": "with_comments",
                        "in": "query",
                        "type": "boolean",
                        "description": "Include the comments.",
                        "default": false
                    },
                    {
                        "name": "with_bundles",
                        "in": "query",
                        "type": "boolean",
                        "description": "Include the bundles.",
                        "default": false
                    },
                    {
                        "name": "with_sightings",
                        "in": "query",
                        "type": "boolean",
                        "description": "Include the sightings.",
                        "default": false
                    }
                ],
                "tags": [
                    "vulnerability"
                ]
            }
        }
    },
    "info": {
        "title": "Vulnerability-Lookup API",
        "version": "2.10.0",
        "description": "<a href='https:\/\/github.com\/vulnerability-lookup\/vulnerability-lookup' rel='noreferrer' target='_blank'><img src='https:\/\/vulnerability.circl.lu\/static\/img\/VL-hori-coul.png' width='500px' \/><\/a><br \/>API to query <a href='https:\/\/github.com\/vulnerability-lookup\/vulnerability-lookup' rel='noreferrer' target='_blank'>Vulnerability-Lookup<\/a>. A project from <a href='https:\/\/www.circl.lu' target='_blank'>CIRCL<\/a>.<br \/><br \/>Back to the <a href='\/'>main page<\/a><br \/><br \/><a href='https:\/\/www.vulnerability-lookup.org\/documentation' target='_blank'>Official documentation<\/a> of the project.",
        "license": {
            "name": "GNU Affero General Public License version 3",
            "url": "https:\/\/www.gnu.org\/licenses\/agpl-3.0.html"
        }
    },
    "produces": [
        "application\/json"
    ],
    "consumes": [
        "application\/json"
    ],
    "securityDefinitions": {
        "apikey": {
            "type": "apiKey",
            "in": "header",
            "name": "X-API-KEY"
        }
    },
    "security": [
        {
            "apikey": []
        }
    ],
...
}